IngredientesDAO

+++ src/java/Modelo/dto/RegionDTO.java
@@ -7,13 +7,15 @@ import java.util.Objects;
 
 public class RegionDTO implements Serializable {
      
-    private float codigo ;
+    private int codigo ;
     private String nombre_region ;
 
-    public RegionDTO(float codigo, String nombre_region) {
+    public RegionDTO(int codigo, String nombre_region) {
         this.codigo = codigo;
         this.nombre_region = nombre_region;
     }
+
+    
 
     public RegionDTO(String nombre_region) {
         this.nombre_region = nombre_region;
@@ -24,11 +26,11 @@ public class RegionDTO implements Serializable {
         
     }   
 
-    public float getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(float codigo) {
+    public void setCodigo(int codigo) {
         this.codigo = codigo;
     }
 
@@ -39,6 +41,9 @@ public class RegionDTO implements Serializable {
     public void setNombre_region(String nombre_region) {
         this.nombre_region = nombre_region;
     }
+
+   
+    
     
     @Override
     public String toString (){
